requirements.txt
Django==5.1.6
fitz==0.0.1.dev2
pdf2image==1.17.0
Pillow==11.1.0
PyAutoGUI==0.9.54
pynput==1.8.0
pytesseract==0.3.13
Requests==2.32.3
